[agent] test(dashboard): cover exact persistent framing
